Added AbstractChainingSpanProcessor
